separate serverside socket logic to separate file, declare imports for PlayerList.tsx
